- separate `PaperOnboarding` delegate & dataSource from `PaperOnboarding.swift` file - make delegate method optional
