+++ Source/PaperOnboarding.swift
@@ -10,57 +10,6 @@ import UIKit
 
 public typealias OnboardingItemInfo = (imageName: UIImage, title: String, description: String, iconName: UIImage, color: UIColor, titleColor: UIColor, descriptionColor: UIColor, titleFont: UIFont, descriptionFont: UIFont)
 
-/**
- *  The PaperOnboardingDataSource protocol is adopted by an object that mediates the application’s data model for a PaperOnboarding object.
- The data source information it needs to construct and modify a PaperOnboarding.
- */
-public protocol PaperOnboardingDataSource {
-  /**
-   Asks the data source to return the number of items.
-   
-   - parameter index: An index of item in PaperOnboarding.
-   
-   - returns: The number of items in PaperOnboarding.
-   */
-  func onboardingItemsCount() -> Int
-  
-  /**
-   Asks the data source for configureation item.
-   
-   - parameter index: An index of item in PaperOnboarding.
-   
-   - returns: configuration info for item
-   */
-  func onboardingItemAtIndex(_ index: Int) -> OnboardingItemInfo
-}
-
-/**
- *  The delegate of a PaperOnboarding object must adopt the PaperOnboardingDelegate protocol. Optional methods of the
- protocol allow the delegate to manage items, configure items, and perform other actions.
- */
-public protocol PaperOnboardingDelegate {
-  /**
-   Tells the delegate that the paperOnbording start scrolling.
-   
-   - parameter index: An curretn index item
-   */
-  func onboardingWillTransitonToIndex(_ index: Int)
-  
-  /**
-   Tells the delegate that the specified item is now selected
-   
-   - parameter index: An curretn index item
-   */
-  func onboardingDidTransitonToIndex(_ index: Int)
-  
-  /**
-   Tells the delegate the PaperOnboarding is about to draw a item for a particular row. Use this method for configure items
-   
-   - parameter item:  A OnboardingContentViewItem object that PaperOnboarding is going to use when drawing the row.
-   - parameter index: An curretn index item
-   */
-  func onboardingConfigurationItem(_ item: OnboardingContentViewItem, index: Int)
-}
 
 ///An instance of PaperOnboarding which display collection of information.
 open class PaperOnboarding: UIView {
@@ -108,7 +57,6 @@ open class PaperOnboarding: UIView {
 }
 
 // MARK: methods
-
 public extension PaperOnboarding {
   
   /**
@@ -139,8 +87,8 @@ public extension PaperOnboarding {
     
   }
 }
-// MARK: create
 
+// MARK: create
 extension PaperOnboarding {
   
   fileprivate func commonInit() {
@@ -189,7 +137,6 @@ extension PaperOnboarding {
 }
 
 // MARK: helpers
-
 extension PaperOnboarding {
   
   fileprivate func backgroundColor(_ index: Int) -> UIColor {
@@ -201,7 +148,6 @@ extension PaperOnboarding {
 }
 
 // MARK: GestureControlDelegate
-
 extension PaperOnboarding: GestureControlDelegate {
   
   func gestureControlDidSwipe(_ direction: UISwipeGestureRecognizerDirection) {
@@ -217,7 +163,6 @@ extension PaperOnboarding: GestureControlDelegate {
 }
 
 // MARK: OnboardingDelegate
-
 extension PaperOnboarding: OnboardingContentViewDelegate {
   
   func onboardingItemAtIndex(_ index: Int) -> OnboardingItemInfo? {
